:bookmark: Releasing / Version tags. 3.7.0-JDK8

+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/service/impl/SysMenuServiceImpl.java
@@ -1,17 +1,20 @@
 /*
- * Copyright (c) 2020 pig4cloud Authors. All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *      Copyright (c) 2018-2025, lengleng All rights reserved.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions are met:
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Redistributions of source code must retain the above copyright notice,
+ *  this list of conditions and the following disclaimer.
+ *  Redistributions in binary form must reproduce the above copyright
+ *  notice, this list of conditions and the following disclaimer in the
+ *  documentation and/or other materials provided with the distribution.
+ *  Neither the name of the pig4cloud.com developer nor the names of its
+ *  contributors may be used to endorse or promote products derived from
+ *  this software without specific prior written permission.
+ *  Author: lengleng (wangiegie@gmail.com)
+ *
  */
 
 package com.pig4cloud.pig.admin.service.impl;
@@ -20,6 +23,7 @@ import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.lang.tree.Tree;
 import cn.hutool.core.lang.tree.TreeNode;
 import cn.hutool.core.lang.tree.TreeUtil;
+import cn.hutool.core.util.BooleanUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -33,19 +37,22 @@ import com.pig4cloud.pig.common.core.constant.CommonConstants;
 import com.pig4cloud.pig.common.core.constant.enums.MenuTypeEnum;
 import com.pig4cloud.pig.common.core.exception.ErrorCodes;
 import com.pig4cloud.pig.common.core.util.MsgUtils;
-import lombok.RequiredArgsConstructor;
+import com.pig4cloud.pig.common.core.util.R;
+import javax.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.BeanUtils;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.Assert;
 
-import javax.validation.constraints.NotNull;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
@@ -57,34 +64,31 @@ import java.util.stream.Collectors;
  * @since 2017-10-29
  */
 @Service
-@RequiredArgsConstructor
+@AllArgsConstructor
+@Slf4j
 public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> implements SysMenuService {
 
 	private final SysRoleMenuMapper sysRoleMenuMapper;
 
 	@Override
-	@Cacheable(value = CacheConstants.MENU_DETAILS, key = "#roleId  + '_menu'", unless = "#result == null")
-	public Set<SysMenu> findMenuByRoleId(Long roleId) {
+	@Cacheable(value = CacheConstants.MENU_DETAILS, key = "#roleId", unless = "#result.isEmpty()")
+	public List<SysMenu> findMenuByRoleId(Long roleId) {
 		return baseMapper.listMenusByRoleId(roleId);
 	}
 
-	/**
-	 * 级联删除菜单
-	 * @param id 菜单ID
-	 * @return true成功, false失败
-	 */
 	@Override
 	@Transactional(rollbackFor = Exception.class)
 	@CacheEvict(value = CacheConstants.MENU_DETAILS, allEntries = true)
-	public Boolean removeMenuById(Long id) {
+	public R removeMenuById(Long id) {
 		// 查询父节点为当前节点的节点
 		List<SysMenu> menuList = this.list(Wrappers.<SysMenu>query().lambda().eq(SysMenu::getParentId, id));
-
-		Assert.isTrue(CollUtil.isEmpty(menuList), MsgUtils.getMessage(ErrorCodes.SYS_MENU_DELETE_EXISTING));
+		if (CollUtil.isNotEmpty(menuList)) {
+			return R.failed(MsgUtils.getMessage(ErrorCodes.SYS_MENU_DELETE_EXISTING));
+		}
 
 		sysRoleMenuMapper.delete(Wrappers.<SysRoleMenu>query().lambda().eq(SysRoleMenu::getMenuId, id));
 		// 删除当前菜单及其子菜单
-		return this.removeById(id);
+		return R.ok(this.removeById(id));
 	}
 
 	@Override
@@ -95,30 +99,32 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 
 	/**
 	 * 构建树查询 1. 不是懒加载情况，查询全部 2. 是懒加载，根据parentId 查询 2.1 父节点为空，则查询ID -1
-	 * @param lazy 是否是懒加载
 	 * @param parentId 父节点ID
+	 * @param menuName 菜单名称
 	 * @return
 	 */
 	@Override
-	public List<Tree<Long>> treeMenu(boolean lazy, Long parentId) {
-		if (!lazy) {
-			List<TreeNode<Long>> collect = baseMapper
-				.selectList(Wrappers.<SysMenu>lambdaQuery().orderByAsc(SysMenu::getSortOrder))
-				.stream()
-				.map(getNodeFunction())
-				.collect(Collectors.toList());
-
-			return TreeUtil.build(collect, CommonConstants.MENU_TREE_ROOT_ID);
-		}
-
+	public List<Tree<Long>> treeMenu(Long parentId, String menuName, String type) {
 		Long parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
 
 		List<TreeNode<Long>> collect = baseMapper
-			.selectList(
-					Wrappers.<SysMenu>lambdaQuery().eq(SysMenu::getParentId, parent).orderByAsc(SysMenu::getSortOrder))
+			.selectList(Wrappers.<SysMenu>lambdaQuery()
+				.like(StrUtil.isNotBlank(menuName), SysMenu::getName, menuName)
+				.eq(StrUtil.isNotBlank(type), SysMenu::getMenuType, type)
+				.orderByAsc(SysMenu::getSortOrder))
 			.stream()
 			.map(getNodeFunction())
 			.collect(Collectors.toList());
+
+		// 模糊查询 不组装树结构 直接返回 表格方便编辑
+		if (StrUtil.isNotBlank(menuName)) {
+			return collect.stream().map(node -> {
+				Tree<Long> tree = new Tree<>();
+				tree.putAll(node.getExtra());
+				BeanUtils.copyProperties(node, tree);
+				return tree;
+			}).collect(Collectors.toList());
+		}
 
 		return TreeUtil.build(collect, parent);
 	}
@@ -126,24 +132,19 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 	/**
 	 * 查询菜单
 	 * @param all 全部菜单
+	 * @param type 类型
 	 * @param parentId 父节点ID
 	 * @return
 	 */
 	@Override
-	public List<Tree<Long>> filterMenu(Set<SysMenu> all, Long parentId) {
+	public List<Tree<Long>> filterMenu(Set<SysMenu> all, String type, Long parentId) {
 		List<TreeNode<Long>> collect = all.stream()
-			.filter(menu -> MenuTypeEnum.LEFT_MENU.getType().equals(menu.getType()))
-			.filter(menu -> StrUtil.isNotBlank(menu.getPath()))
+			.filter(menuTypePredicate(type))
 			.map(getNodeFunction())
 			.collect(Collectors.toList());
+
 		Long parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
 		return TreeUtil.build(collect, parent);
-	}
-
-	@Override
-	@CacheEvict(value = CacheConstants.MENU_DETAILS, allEntries = true)
-	public void clearMenuCache() {
-
 	}
 
 	@NotNull
@@ -156,15 +157,41 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 			node.setWeight(menu.getSortOrder());
 			// 扩展属性
 			Map<String, Object> extra = new HashMap<>();
-			extra.put("icon", menu.getIcon());
 			extra.put("path", menu.getPath());
-			extra.put("type", menu.getType());
+			extra.put("menuType", menu.getMenuType());
 			extra.put("permission", menu.getPermission());
-			extra.put("label", menu.getName());
 			extra.put("sortOrder", menu.getSortOrder());
-			extra.put("keepAlive", menu.getKeepAlive());
+
+			// 适配 vue3
+			Map<String, Object> meta = new HashMap<>();
+			meta.put("title", menu.getName());
+			meta.put("isLink", menu.getPath() != null && menu.getPath().startsWith("http") ? menu.getPath() : "");
+			meta.put("isHide", !BooleanUtil.toBooleanObject(menu.getVisible()));
+			meta.put("isKeepAlive", BooleanUtil.toBooleanObject(menu.getKeepAlive()));
+			meta.put("isAffix", false);
+			meta.put("isIframe", BooleanUtil.toBooleanObject(menu.getEmbedded()));
+			meta.put("icon", menu.getIcon());
+			// 增加英文
+			meta.put("enName", menu.getEnName());
+
+			extra.put("meta", meta);
 			node.setExtra(extra);
 			return node;
+		};
+	}
+
+	/**
+	 * menu 类型断言
+	 * @param type 类型
+	 * @return Predicate
+	 */
+	private Predicate<SysMenu> menuTypePredicate(String type) {
+		return vo -> {
+			if (MenuTypeEnum.TOP_MENU.getDescription().equals(type)) {
+				return MenuTypeEnum.TOP_MENU.getType().equals(vo.getMenuType());
+			}
+			// 其他查询 左侧 + 顶部
+			return !MenuTypeEnum.BUTTON.getType().equals(vo.getMenuType());
 		};
 	}
